+++ bakedanuki/bakedanuki-util/python/bd_util/_test/test.py
@@ -11,19 +11,17 @@ def main():
     nodes = bdu.Nodes()
 
     trsf = nodes.create.transform(name="test")
-    keys = [(i, i) for i in range(10000)]
 
-    trsf.tx.keyframe.set_keys(keys)
-    trsf.ty.keyframe.set_keys(keys)
-    trsf.tz.keyframe.set_keys(keys)
-
-    trsf.rx.keyframe.set_keys(keys)
-    trsf.ry.keyframe.set_keys(keys)
-    trsf.rz.keyframe.set_keys(keys)
-
-    trsf.sx.keyframe.set_keys(keys)
-    trsf.sy.keyframe.set_keys(keys)
-    trsf.sz.keyframe.set_keys(keys)
+    trsf.tx.keyframe.set_key(-100, frame=-100, out_tangent_type="flat")
+    trsf.tx.keyframe.set_key(100, frame=100, in_tangent_type="flat")
 
     nodes.modifier_manager.do_it_dag()
+    nodes.modifier_manager.do_it_dg()
+
+    samples = trsf.tx.sample_values(frames=range(-100, 101))
+
+    trsf.ty.keyframe.set_keys(
+        samples, in_tangent_type="auto", out_tangent_type="auto"
+    )
+
     nodes.modifier_manager.do_it_dg()